Split #1018 recovery/spare API tests into a dedicated module
Move the range-recovery and spare-pool API flow tests out of
ctf/test_api_view_flows.py (which crossed the 800-line per-module cap
enforced by test_test_suite_structure.py) into
ctf/test_range_recovery_api.py, keeping each test module behavior-scoped.
